app/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
fastapi==0.112.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
python-dotenv==1.0.1
redis==5.0.8
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.5
rodi==2.0.6
asyncclick~=8.1.7.2
PyYAML~=6.0.2